recycle_assistant_project/requirements.txt
streamlit
tensorflow
keras
numpy
pandas
Pillow
plotly

gdown
